fix: adjust search on cmc
