+++ packages/web3-providers/src/cmc/index.ts
@@ -7,39 +7,10 @@ import { BTC_FIRST_LEGER_DATE, CMC_STATIC_BASE_URL, CMC_V1_BASE_URL, THIRD_PARTY
 import { resolveCoinMarketCapChainId } from './helpers.js'
 import { fetchJSON } from '../helpers.js'
 import { getCommunityLink, isMirroredKeyword } from '../trending/helpers.js'
-import type { Coin, ResultData, Status } from './types.js'
+import type { Coin, Pair, ResultData, Status, QuotesInfo, CoinInfo } from './types.js'
 import { FuseTrendingAPI } from '../fuse/index.js'
 
 // #regin get quote info
-export interface QuotesInfo {
-    circulating_supply: number
-    cmc_rank: number
-    date_added: string
-    id: number
-    is_active: boolean
-    is_fiat: 0 | 1
-    last_updated: string
-    max_supply: null | number
-    name: string
-    num_market_pairs: number
-    quote: Record<
-        string,
-        {
-            last_updated: string
-            market_cap?: number
-            percent_change_1h?: number
-            percent_change_7d?: number
-            percent_change_24h?: number
-            price: number
-            volume_24h?: number
-        }
-    >
-    slug: string
-    symbol: string
-    tags: string[]
-    total_supply: number
-}
-
 export async function getQuotesInfo(id: string, currency: string) {
     const params = new URLSearchParams()
     params.append('id', id)
@@ -62,59 +33,6 @@ export async function getQuotesInfo(id: string, currency: string) {
 // #endregion
 
 // #region get coin info
-export interface PlatformInfo {
-    coin: {
-        id: string
-        name: string
-        slug: string
-        symbol: string
-    }
-    name: string
-}
-
-export interface ContractInfo {
-    contract_address: string
-    platform: PlatformInfo
-}
-export interface CoinInfo {
-    category: string
-    contract_address: ContractInfo[]
-    date_added: string
-    date_launched: string | null
-    description: string
-    id: number
-    is_hidden: 0 | 1
-    logo: string
-    name: string
-    notice: string
-    platform?: {
-        id: number
-        name: string
-        slug: string
-        symbol: string
-        token_address: string
-    }
-    slug: string
-    status: string
-    subreddit: string
-    symbol: string
-    'tag-groups': string[]
-    'tag-names': string[]
-    tags: string[]
-    twitter_username: string
-    urls: {
-        announcement?: string[]
-        chat?: string[]
-        explorer?: string[]
-        reddit?: string[]
-        source_code?: string[]
-        message_board?: string[]
-        technical_doc?: string[]
-        twitter?: string[]
-        website?: string[]
-    }
-}
-
 export async function getCoinInfo(id: string) {
     const params = new URLSearchParams('aux=urls,logo,description,tags,platform,date_added,notice,status')
     params.append('id', id)
@@ -135,48 +53,7 @@ export async function getCoinInfo(id: string) {
 // #endregion
 
 // #region latest market pairs
-export interface Pair {
-    exchange: {
-        id: number
-        name: string
-        slug: string
-    }
-    market_id: number
-    market_pair: string
-    market_pair_base: {
-        currency_id: number
-        currency_symbol: string
-        currency_type: string
-        exchange_symbol: string
-    }
-    market_pair_quote: {
-        currency_id: number
-        currency_symbol: string
-        currency_type: string
-        exchange_symbol: string
-    }
-    market_reputation: number
-    market_score: number
-    market_url: string
-    outlier_detected: 0 | 1
-    quote: Record<
-        string,
-        {
-            effective_liquidity: 0 | 1
-            last_updated: string
-            price: number
-            price_quote: number
-            volume_24h: number
-        }
-    > & {
-        exchange_reported: {
-            last_updated: string
-            price: number
-            volume_24h_base: number
-            volume_24h_quote: number
-        }
-    }
-}
+
 export async function getLatestMarketPairs(id: string, currency: string) {
     const params = new URLSearchParams(
         'aux=num_market_pairs,market_url,price_quote,effective_liquidity,market_score,market_reputation&limit=40&sort=cmc_rank&start=1',
@@ -250,13 +127,17 @@ export class CoinMarketCapAPI implements TrendingAPI.Provider<ChainId> {
                 name: x.name,
                 symbol: x.symbol,
                 type: TokenType.Fungible,
-                contract_address: x.platform?.name === 'Ethereum' ? x.platform.token_address : undefined,
+                contract_address: x.platform?.token_address,
+                market_cap_rank: x.rank,
             }))
     }
 
     async getCoinsByKeyword(chainId: ChainId, keyword: string): Promise<TrendingAPI.Coin[]> {
         const coins = await this.fuse.getSearchableItems(this.getAllCoins)
-        return coins.search(keyword).map((x) => x.item)
+        return coins
+            .search(keyword)
+            .map((x) => x.item)
+            .slice(0, 10)
     }
 
     async getCoinTrending(chainId: ChainId, id: string, currency: TrendingAPI.Currency): Promise<TrendingAPI.Trending> {
